# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Avia-back API",
        "version": "1.0.3"
    },
    "paths": {
...
        "/api/v1/admin/documents": {
            "get": {
                "tags": [
                    "Документы"
                ],
                "summary": "Получить список всех документов",
                "description": "Для супер-администратора возвращает список всех документов вместе со списком соответствующих технических систем, содержащихся в системе. Для администратора возвращает список только тех документов (включая список соответствующих технических систем) на технические системы, которые доступны в рамках проекта для той организации к которой принадлежит администратор.",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "Some code"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "type": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "version": {
                                                "type": "string",
                                                "example": "Some version"
                                            },
                                            "file": {
                                                "type": "string",
                                                "example": "Some path to file"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-09-15T01:52:11.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2023-09-15T01:52:11.000000Z"
                                            },
                                            "technical_systems": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "code": {
                                                            "type": "string",
                                                            "example": "Some code"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Some name"
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Some description"
                                                        },
                                                        "parent_technical_system_id": {
                                                            "type": "integer",
                                                            "example": null
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2023-09-15T01:52:11.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Документы"
                ],
                "summary": "Создание нового документа",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "code": {
                                                "type": "string",
                                                "example": "Some code"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Some name"
                                            },
                                            "type": {
                                                "type": "integer",
                                                "example": 0
                                            },
                                            "version": {
                                                "type": "string",
                                                "example": "Some version"
                                            },
                                            "file": {
                                                "type": "string",
                                                "example": "Some path to file"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "Some code"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Some name"
                                        },
                                        "type": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "version": {
                                            "type": "string",
                                            "example": "Some version"
                                        },
                                        "file": {
                                            "type": "string",
                                            "example": "Some path to file"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-15T01:52:11.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-09-15T01:52:11.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}